refactor: move Products dto to application and domain folders
