+++ src/application/dtos/product/index.ts
@@ -1,1 +1,2 @@
+export * from './product-repository-output'
 export * from './products-repository-output'
