+++ src/application/dtos/product/product-repository-output.ts
@@ -1,0 +1,3 @@
+import { type ProductOutput } from '@/domain/dtos/product'
+
+export type ProductRepositoryOutput = ProductOutput
