+++ src/application/protocols/product/queries/get-latest-products-repository.ts
@@ -1,4 +1,4 @@
-import { type ProductsRepositoryOutput } from '@/application/protocols/product/common'
+import { type ProductsRepositoryOutput } from '@/application/dtos/product'
 
 export interface GetLatestProductsRepository {
   getLatest: (page: number, limit: number) => Promise<ProductsRepositoryOutput>
