+++ src/application/protocols/product/queries/get-products-by-category-repository.ts
@@ -1,4 +1,4 @@
-import { type ProductsRepositoryOutput } from '@/application/protocols/product/common'
+import { type ProductsRepositoryOutput } from '@/application/dtos/product'
 
 export interface GetProductsByCategoryRepository {
   getByCategory: (category: string, page: number, limit: number) => Promise<ProductsRepositoryOutput>
