+++ src/domain/dtos/product/index.ts
