+++ src/domain/dtos/product/product-output.ts
