+++ src/domain/usecases/product/queries/get-latest-products.ts
@@ -1,4 +1,4 @@
-import { type ProductsOutput } from '@/domain/entities/product/dto'
+import { type ProductsOutput } from '@/domain/dtos/product'
 
 export interface GetLatestProducts {
   getLatest: (page: number, limit: number) => Promise<ProductsOutput>
