+++ src/domain/usecases/product/queries/get-product-by-slug.ts
@@ -1,4 +1,4 @@
-import { type ProductOutput } from '@/domain/entities/product/dto'
+import { type ProductOutput } from '@/domain/dtos/product'
 
 export interface GetProductBySlug {
   getBySlug: (slug: string) => Promise<ProductOutput>
