+++ src/domain/usecases/product/queries/get-products-by-category.ts
@@ -1,4 +1,4 @@
-import { type ProductsOutput } from '@/domain/entities/product/dto'
+import { type ProductsOutput } from '@/domain/dtos/product'
 
 export interface GetProductsByCategory {
   getByCategory: (category: string, page: number, limit: number) => Promise<ProductsOutput>
