+++ src/domain/usecases/product/queries/get-products-with-discount.ts
@@ -1,4 +1,4 @@
-import { type ProductsOutput } from '@/domain/entities/product/dto'
+import { type ProductsOutput } from '@/domain/dtos/product'
 
 export interface GetProductsWithDiscount {
   getWithDiscount: (page: number, limit: number) => Promise<ProductsOutput>
