+++ src/infra/db/mongodb/product/queries/get-products-with-discount-mongo-repository.ts
@@ -1,6 +1,6 @@
 import { MongoHelper } from '@/infra/db/mongodb/helpers'
 import { type GetProductsWithDiscountRepository } from '@/application/protocols/product/queries'
-import { type ProductsRepositoryOutput } from '@/application/protocols/product/common'
+import { type ProductsRepositoryOutput } from '@/application/dtos/product'
 
 export class GetProductsWithDiscountMongoRepository implements GetProductsWithDiscountRepository {
   private readonly mongoHelper: MongoHelper = MongoHelper.getInstance()
