+++ tests/domain/mocks/product/mock-product.ts
@@ -2,7 +2,7 @@ import { faker } from '@faker-js/faker'
 
 import { type AddProduct, type UpdateProduct } from '@/domain/usecases/product/commands'
 import { ProductHelper } from '@/domain/entities/product'
-import { type ProductOutput, type ProductsOutput } from '@/domain/entities/product/dto'
+import { type ProductOutput, type ProductsOutput } from '@/domain/dtos/product'
 
 export const mockAddProductInput = (): AddProduct.Input => ({
   name: faker.string.alpha({ length: { min: 3, max: 20 } }),
